# TPBookstore-api/config/swagger.yaml
openapi: 3.0.1
info:
  title: npsstore API
  description: This is a Backend API of NBSStore
  version: 1.0.0
  termsOfService: http://quocbaoit.com/terms/
  contact:
    email: NBSstoreProject@quocbaoit.com
...
paths:
...
  /api/v1/user:
    post:
      summary: Create new user account
      description: This can only be done by the logged in user.
      tags:
        - User
      requestBody:
        description: Payload to create new user
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  nullable: false
                email:
                  type: string
                  nullable: false
                password:
                  type: string
                  nullable: false
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/MyUser'
                  - type: object
                    properties:
                      refreshToken:
                        type: string
                        description: JWT Refresh Token
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Get all users by admin
      description: Get all users by admin
      tags:
        - User
      parameters:
        - in: query
          name: dateOrder
          required: false 
          schema:
            type: string
          description: Sort users by date (newest/latest)
        - in: query
          name: status
          required: false
          schema: 
            type: string
          description: Filter users by status (disabled/notDisabled/all)        
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/MyUser'
                    - type: object
                      properties:
                        createdAt:
                          type: string
                          description: The date and time when this record is created
                          format: date-time
                        updatedAt:
                          type: string
                          description: The date and time when this record is updated
                          format: date-time
                        __v:
                          type: number
                          description: Document version key
                        password:
                          type: string
                          description: User password
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/user/{id}:
    delete:
      summary: Delete existed user
      description: Delete existed user
      security:
        - bearerAuth: []
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    MyUser:
      type: object
      properties:
        _id:
          type: string
          description: user id
        name:
          type: string
          description: username
        email:
          type: string
          description: email
        avatarUrl:
          type: string
          description: avatar url
        isAdmin:
          type: boolean
          description: is Admin or not
        token:
          type: string
          description: JWT login token
        isDisabled:
          type: boolean
          description: is Disabled or not
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: error message
        stack:
          type: string
          description: error message description
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: apiKey
      description: "<i>JWT Authorization header using the Bearer scheme.</i> <b>Example</b>: <i><u>Bearer 12345abcdef</u></i>"
      in: header          # can be "header", "query" or "cookie"
      name: Authorization # name of the header, query parameter or cookie
# 2) Apply the security globally to all operations
security:
  - bearerAuth: [] # use the same name as above